requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-bootstrap3==23.4
django-filter==23.2
django-tables2==2.6.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
logging==0.4.9.6
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.2.0
packaging==23.1
pathlib==1.0.1
pluggy==1.2.0
py==1.11.0
Pygments==2.15.1
PySocks==1.7.1
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tenacity==8.2.2
trio==0.22.1
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0
